+++ packages/apptive_grid_core/test/grid_field_test.dart
@@ -25,4 +25,33 @@ void main() {
       expect(a.hashCode, isNot(b.hashCode));
     });
   });
+
+  group('Json Parsing', () {
+    test('Parses from and to json', () {
+      final json = {
+        "type": {
+          "name": "enum",
+          "options": ["a", "b"],
+          "componentTypes": ["selectBox"]
+        },
+        "key": "keyTest",
+        "name": "name",
+        "schema": {
+          "type": "string",
+          "enum": ["a", "b"]
+        },
+        "id": "ep8m43a1i4v1lgj4q8r31ba2f",
+        "_links": {}
+      };
+
+      final fromJson = GridField.fromJson(json);
+
+      expect(fromJson.id, 'ep8m43a1i4v1lgj4q8r31ba2f');
+      expect(fromJson.name, 'name');
+      expect(fromJson.key, 'keyTest');
+      expect(fromJson.type, DataType.singleSelect);
+
+      expect(GridField.fromJson(fromJson.toJson()), equals(fromJson));
+    });
+  });
 }